Create R-type Machine Code
So I was trying to print the array just to make sure the declaration in the .data was correct but it did not build. Gotta find out what data type and what format it needs to be. Going to take the instruction from Caseys program and compare it to the strings in the array. When the string has found it's match we will have the funct binary and then I'll have to follow up with the register compares. Next will be finding a way to handle immediate values....